renamed attachDisk and detachDisk, removed interfaces input in attachDisk and moved attachDisk and detachDisk in disks folder
